Changed table types, to better match rails' ones (float, boolean, decimal).

+++ test/test_app/db/migrate/20090208151615_create_products.php
@@ -7,7 +7,7 @@ class CreateProduct extends ActiveRecord_Migration
     $t = $this->db->new_table('products');
     
     $t->add_column('name',  'string', array('null' => false, 'limit' => 100));
-    $t->add_column('price', 'double', array('null' => false, 'signed' => false));
+    $t->add_column('price', 'float',  array('null' => false, 'signed' => false));
     $t->add_timestamps();
     
     return $t->create();
